Windows 10 Support + Error Message + Custom Icon
Added support for Windows 10, which means you can create the server on any Windows;
Added a fake Error Message, saying that you have low Virtual Memory. It is commented out at the very end of code, if you want to activate it, uncomment it.
Added possibility for custom icon. All you have to do is place the icon in the same directory as this script and rename it "icon.ico"